% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collabExample.R
\docType{package}
\name{collabExample-package}
\alias{collabExample}
\alias{collabExample-package}
\title{collabExample: R package by collaboration}
\description{
An example package for collaboartion experiance.
}
\details{
See `vignette("collabExample", package = "collabExample")` for an overview
of the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sktivd/collab_example}
  \item Report bugs at \url{https://github.com/sktivd/collab_example/issues}
}

}
\author{
\strong{Maintainer}: Sung Gon Yi \email{skonmeme@gmail.com}

Authors:
\itemize{
  \item Sangjo Han \email{sangjo.han@gmail.com}
  \item Taegyun Yoon \email{bacillus@sk.com}
  \item Junghyun Namkung \email{jh.namkung@gmail.com}
  \item Yonghwan Choi \email{choiyh1976@gmail.com}
  \item Jungsun Park \email{pj0518@gmail.com}
}

}
