% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sangjohan.R
\name{eatSomething}
\alias{eatSomething}
\title{eatSomething}
\usage{
eatSomething(food)
}
\arguments{
\item{food}{food name for me to eat}
}
\value{
I am now eating some string to \code{food}
}
\description{
eatSomething
}
\examples{
eatSomething('Noodle')
}
