% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSort.R
\name{mergeSort}
\alias{mergeSort}
\title{Merge Sort}
\usage{
mergeSort(unsorted, ascending = TRUE)
}
\arguments{
\item{unsorted}{a numeric array}

\item{ascending}{sorted by ascending value (default: TRUE)}
}
\value{
sorted numeric array
}
\description{
sort a numeric array by merge sort algorithm
}
\examples{
mergeSort(c(324, 234, 243))
}
