% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\docType{data}
\name{people}
\alias{people}
\title{People in SKT IVD organization}
\format{A \code{data.frame} with 3 columns. The columns are defined as follows:
\describe{
 \item{\code{ID}}{GitHub ID}
 \item{\code{Name}}{Full name}
 \item{\code{Email}}{E-mail address}
}}
\source{
Joined member list in SKT IVD organization
}
\usage{
people
}
\description{
This is a list of SKT IVD organization members joined in GitHub
}
\examples{
data(people)
}
\keyword{datasets}
