% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayHello.R
\name{sayHello}
\alias{sayHello}
\title{Say Hello}
\usage{
sayHello(whom)
}
\arguments{
\item{whom}{name to say hello}
}
\value{
the say hello string to \code{whom}
}
\description{
Say Hello
}
\examples{
sayHello('skon')
}
