% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacillust.R
\name{sayHelloRepeat}
\alias{sayHelloRepeat}
\title{Say Hello}
\usage{
sayHelloRepeat(iter)
}
\arguments{
\item{iter}{num of say hello}
}
\value{
the say string with iter nub
}
\description{
Say Hello with repeatition
}
\examples{
sayHelloRepeat(3)
}
