% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacillust.R
\name{sayHelloRepeat}
\alias{sayHelloRepeat}
\title{Say Hello}
\usage{
sayHelloRepeat(iter)
}
\arguments{
\item{iter}{num of say hello}

\item{whom}{name to say hello}
}
\value{
the say hello string to \code{whom}
Say Hello with repeatition
}
\description{
Say Hello
}
\examples{
sayHelloRepeat(3)
}
